feat: move interface tests into the correct folders
