+++ src/test/interfaces/database/beta/filters.ts
@@ -1,6 +1,6 @@
 import { Database } from '@ajs/database/beta';
 import { expect } from 'chai';
-import { getUniqueUsers, User } from '../datasets/users';
+import { getUniqueUsers, User } from '../../../datasets/users';
 
 const db = Database<{ [table]: User }>('test-filters');
