+++ src/test/interfaces/database/beta/merge_operations.ts
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/no-unused-expressions */
 import { Database } from '@ajs/database/beta';
 import { expect } from 'chai';
-import { getUniqueUsers, User } from '../datasets/users';
-import { getUniqueProducts, Product } from '../datasets/products';
-import { getUniqueOrders, Order } from '../datasets/orders';
+import { getUniqueUsers, User } from '../../../datasets/users';
+import { getUniqueProducts, Product } from '../../../datasets/products';
+import { getUniqueOrders, Order } from '../../../datasets/orders';
 
 const db = Database<{
   [ordersTable]: Order;
